+++ src/common/Widgets/GTableLayout.cpp
@@ -229,7 +229,7 @@ DistributeSize(	GArray<int> &a,
 	if (Fill.Length())
 	{
 		// Distribute size amongst the growable cells
-		int PerFillSize = AdditionalSize / Fill.Length();
+		int PerFillSize = AdditionalSize / (int)Fill.Length();
 		if (PerFillSize <= 0)
 		    PerFillSize = 1;
 		    
@@ -252,7 +252,7 @@ DistributeSize(	GArray<int> &a,
 			if (a[Cell] && ExistingGrowPx)
 			    Add = a[Cell] * AdditionalSize / ExistingGrowPx;
 			else
-			    Add = MAX(1, AdditionalSize / Grow.Length());
+			    Add = (int)MAX(1, AdditionalSize / Grow.Length());
 			a[Cell] = a[Cell] + Add;
 		}
 	}
@@ -267,7 +267,7 @@ DistributeSize(	GArray<int> &a,
 			if (a[Cell] && UnknownPx)
 			    Add = a[Cell] * AdditionalSize / UnknownPx;
 			else
-			    Add = MAX(1, AdditionalSize / Unknown.Length());
+			    Add = (int)MAX(1, AdditionalSize / Unknown.Length());
 			a[Cell] = a[Cell] + Add;
 		}
 	}
@@ -1650,7 +1650,7 @@ void GTableLayoutPrivate::LayoutVertical(GRect &Client, int *MinY, int *MaxY, Ce
 						if (MinY > InitMinY)
 						{
 							// Allocate any extra min px somewhere..
-							int Amt = (MinY - InitMinY) / AddTo.Length();
+							int Amt = (MinY - InitMinY) / (int)AddTo.Length();
 							for (int i=0; i<AddTo.Length(); i++)
 							{
 								int Idx = AddTo[i];
@@ -1661,7 +1661,7 @@ void GTableLayoutPrivate::LayoutVertical(GRect &Client, int *MinY, int *MaxY, Ce
 						if (MaxY > InitMaxY)
 						{
 							// Allocate any extra max px somewhere..
-							int Amt = (MaxY - InitMaxY) / AddTo.Length();
+							int Amt = (MaxY - InitMaxY) / (int)AddTo.Length();
 							for (int i=0; i<AddTo.Length(); i++)
 							{
 								int Idx = AddTo[i];
@@ -2166,7 +2166,7 @@ void GTableLayout::Value(int64 v)
     GArray<GRadioButton*> Btns;
     if (d->CollectRadioButtons(Btns))
     {
-        if (v >= 0 && v < Btns.Length())
+        if (v >= 0 && v < (ssize_t)Btns.Length())
             Btns[(int)v]->Value(true);
     }
 }
